requirements.txt
streamlit
langchain
openai
python-dotenv
python-pptx
requests
fastmcp
duckduckgo-search